Arreglando problema con la base de datos, mas ejemplo de seeder

+++ database/migrations/2017_04_19_062821_create_clients_table.php
@@ -14,7 +14,8 @@ class CreateClientsTable extends Migration
     public function up()
     {
         Schema::create('clients', function (Blueprint $table) {
-            $table->float('ci_cliente');
+           // $table->float('ci_cliente');
+            $table->increments('ci_cliente');
             $table->string('nombre');
             $table->string('apellido');
             $table->float('telefono');
